Move cart_modal to cart/show, delete unused references to modules and excessive cache

+++ db/migrate/20210831124232_add_order_to_cart_items.rb
@@ -1,6 +1,6 @@
 class AddOrderToCartItems < ActiveRecord::Migration[6.1]
   def change
     add_reference :cart_items, :order, null: true, foreign_key: true
-    change_column :cart_items, :cart_id, :integer, null: true
+    change_column :cart_items, :cart_id, :bigint, null: true
   end
 end
